requirements.txt
diffusers==0.7.2
transformers==4.24.0
scipy==1.7.3
ftfy==6.1.1
python-telegram-bot>=20.0a2
python-dotenv>=0.20.0
huggingface_hub==0.10.1
